Moving LikelihoodOfSearch
